Merge branch 'tida-eventinfo-fix' into '21.1-dev' -- push to 21.0

See merge request !3526

Former-commit-id: 7248c77d78bc43f852e067c3058bceba63c32131

+++ Trigger/TrigAnalysis/TrigInDetAnalysisExample/TrigInDetAnalysisExample/T_AnalysisConfig_Tier0.h
@@ -366,7 +366,7 @@ protected:
     const xAOD::EventInfo* pEventInfo;
 #endif
     unsigned run_number        = 0;
-    unsigned event_number      = 0;
+    unsigned long long event_number      = 0;
     unsigned lumi_block        = 0;
     unsigned bunch_crossing_id = 0;
     unsigned time_stamp = 0;
